+++ test/bigtest/interactors/line-details-page.js
@@ -6,7 +6,6 @@ import {
 } from '@bigtest/interactor';
 
 import Button from './button';
-import { TIMEOUT } from './const';
 
 @interactor class LineDetailsPageActions {
   static defaultScope = '#data-test-line-details-actions';
@@ -47,10 +46,10 @@ export default interactor(class LineDetailsPage {
 
   actions = new LineDetailsPageActions();
   whenLoaded() {
-    return this.timeout(TIMEOUT).when(() => this.isLoaded);
+    return this.timeout(20000).when(() => this.isLoaded);
   }
 
   whenInvoicesLoaded() {
-    return this.timeout(TIMEOUT).when(() => this.lineInvoicesIsPresent);
+    return this.timeout(20000).when(() => this.lineInvoicesIsPresent);
   }
 });