+++ test/bigtest/interactors/line-edit-page.js
@@ -15,7 +15,6 @@ import {
 import { OptionListInteractor } from '@folio/stripes-acq-components/test/bigtest/interactors';
 
 import Button from './button';
-import { TIMEOUT } from './const';
 import { ACCORDION_ID } from '../../../src/components/POLine/const';
 
 const ITEM_DETAILS = {
@@ -221,6 +220,6 @@ export default interactor(class LineEditPage {
   hasTemplateField = isPresent('[name="template"]');
 
   whenLoaded() {
-    return this.timeout(TIMEOUT).when(() => this.isLoaded);
+    return this.timeout(20000).when(() => this.isLoaded);
   }
 });