+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Stories
 
 - [UIOR-521](https://issues.folio.org/browse/UIOR-521) New icon for "Close order"
+- [UIOR-471](https://issues.folio.org/browse/UIOR-471) Update purchase order "Renewal information" accordion
 - [UIOR-506](https://issues.folio.org/browse/UIOR-506) Replace org selection components with organization lookup
 - [UIOR-491](https://issues.folio.org/browse/UIOR-491) Cloning a PO and its POL(s)
 - [UIOR-457](https://issues.folio.org/browse/UIOR-457) Ability to "Reopen" and once again "Close" purchase orders